groq_medical_app/requirements.txt
Flask
groq
python-dotenv
requests
